refactor : ApplyQueryRepository, FriendQueryRepository 패키지 이동
